(g_flags) changing to int from int8.

+++ src/options.c
@@ -28,7 +28,7 @@
 #include "options.h"
 
 /* Global variables, used by all (or nearly all) functions */
-int8 g_flags = NONE;     /* program options */
+int g_flags = NONE;     /* program options */
 /* char *g_directory = NULL;*/ /* output directory */
 /* FILE *g_file = NULL;*/     /* input file */
